refactor: ListViewReservation들 이름 이상하게 잡힌 것 수정
- props 타이핑
